Remove TicketParser.parseSync() and extractSync()
Although the only thing that's currently asynchronous is the call to
zlib.inflate(), the payload lends itself very well to streaming parsing
in principle.  Maintaining synchronous and aynchronous code paths is not
a problem *now*, but it would become a headache if asynchronous parsing
were to be implemented later on. I'll leave parseSync() out for now.